refactor: 重组组件目录结构 - 第二阶段
- 将基础UI组件移至 components/ui/ 目录
  - MetricCard, CustomResizeHandle, ViewAllLink
- 将业务组件移至 components/business/ 目录
  - Header, WelcomeSection, SettingsMenu
- 将活动组件移至 business/activities/ 目录
  - Activities, ResizableActivities
- 将指标组件移至 business/metrics/ 目录
  - AssetsMetric, ClientsMetric, PlansMetric, TasksMetric
- 将小部件组件移至 business/widgets/ 目录
  - Events, RecentPlans, Tasks, Workflows
- 更新所有受影响的导入路径

按照三层架构（布局层、UI层、业务层）组织代码结构
