+++ src/components/business/activities/ResizableActivities.tsx
