+++ src/components/business/metrics/AssetsMetric.tsx
@@ -1,6 +1,6 @@
 import { DollarSign } from 'lucide-react';
 
-import { MetricCard } from '../MetricCard';
+import { MetricCard } from '../../ui/MetricCard';
 
 export const AssetsMetric = () => {
   return (
