+++ src/components/business/metrics/ClientsMetric.tsx
@@ -1,6 +1,6 @@
 import { Users } from 'lucide-react';
 
-import { MetricCard } from '../MetricCard';
+import { MetricCard } from '../../ui/MetricCard';
 
 export const ClientsMetric = () => {
   return (
