+++ src/components/business/metrics/TasksMetric.tsx
@@ -1,6 +1,6 @@
 import { CheckCircle } from 'lucide-react';
 
-import { MetricCard } from '../MetricCard';
+import { MetricCard } from '../../ui/MetricCard';
 
 export const TasksMetric = () => {
   return (
