+++ src/components/business/widgets/Workflows.tsx
@@ -1,7 +1,7 @@
 import { Button } from '@untitled-ui/components/base/buttons/button';
 import { ArrowUpRight, MoreHorizontal, Plus } from 'lucide-react';
 
-import { ViewAllLink } from '../ui/ViewAllLink';
+import { ViewAllLink } from '../../ui/ViewAllLink';
 
 interface WorkflowItem {
   id: string;
